Finished initial build of final project aka Movie Program

+++ final/FinalProject/Movie1960s.cs
@@ -1,6 +1,6 @@
 public class Movie1960s : Movie{
 
-public Movie1960s(string name, string genre, string director, int rating, int year){
+public Movie1960s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
