+++ final/FinalProject/Movie1970s.cs
@@ -1,6 +1,6 @@
 public class Movie1970s : Movie{
 
-public Movie1970s(string name, string genre, string director, int rating, int year){
+public Movie1970s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
