+++ final/FinalProject/Movie1980s.cs
@@ -1,6 +1,6 @@
 public class Movie1980s : Movie{
 
-public Movie1980s(string name, string genre, string director, int rating, int year){
+public Movie1980s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
