+++ final/FinalProject/Movie1990.cs
@@ -1,6 +1,6 @@
 public class Movie1990s : Movie{
 
-public Movie1990s(string name, string genre, string director, int rating, int year){
+public Movie1990s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
