+++ final/FinalProject/Movie2000s.cs
@@ -1,6 +1,6 @@
 public class Movie2000s : Movie{
 
-public Movie2000s(string name, string genre, string director, int rating, int year){
+public Movie2000s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
