+++ final/FinalProject/Movie2010s.cs
@@ -1,6 +1,6 @@
 public class Movie2010s : Movie{
 
-public Movie2010s(string name, string genre, string director, int rating, int year){
+public Movie2010s(string name, string genre, string director, float rating, int year){
     _name = name;
     _genre = genre;
     _director = director;
